requirements.txt
asgiref==3.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.20.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
virtualenv==20.35.3
whitenoise==6.11.0
